refactored handlers to include boilerplate code at higher level
