+++ 00_Base/src/interfaces/messages/AbstractMessageHandler.ts
@@ -20,6 +20,7 @@ export abstract class AbstractMessageHandler implements IMessageHandler {
      */
 
     protected _config: SystemConfig;
+    protected _module?: IModule;
     protected _logger: Logger<ILogObj>;
 
     /**
@@ -28,9 +29,29 @@ export abstract class AbstractMessageHandler implements IMessageHandler {
      * @param config The system configuration.
      * @param logger [Optional] The logger to use.
      */
-    constructor(config: SystemConfig, logger?: Logger<ILogObj>) {
+    constructor(config: SystemConfig, logger?: Logger<ILogObj>, module?: IModule, ) {
         this._config = config;
+        this._module = module;
         this._logger = logger ? logger.getSubLogger({ name: this.constructor.name }) : new Logger<ILogObj>({ name: this.constructor.name });
+    }
+
+    /**
+     * Getter & Setter
+     */
+
+    get module(): IModule | undefined {
+        return this._module;
+    }
+    set module(value: IModule | undefined) {
+        this._module = value;
+    }
+
+    /**
+     * Methods
+     */
+
+    async handle(message: IMessage<OcppRequest | OcppResponse | OcppError>, props?: HandlerProperties): Promise<void> {
+        await this._module?.handle(message, props);
     }
 
     /**
@@ -39,9 +60,5 @@ export abstract class AbstractMessageHandler implements IMessageHandler {
 
     abstract subscribe(identifier: string, actions?: CallAction[], filter?: { [k: string]: string; }): Promise<boolean>;
     abstract unsubscribe(identifier: string): Promise<boolean>;
-    abstract handle(message: IMessage<OcppRequest | OcppResponse | OcppError>, props?: HandlerProperties): Promise<void>;
     abstract shutdown(): void;
-
-    abstract get module(): IModule | undefined;
-    abstract set module(value: IModule | undefined);
 }
